Fix vrconfig load

+++ server/core/src/main/java/dev/slimevr/config/TrackingChecklistConfig.kt
@@ -1,5 +1,5 @@
 package dev.slimevr.config
 
 class TrackingChecklistConfig {
-	val ignoredStepsIds: MutableList<UByte> = mutableListOf()
+	val ignoredStepsIds: MutableList<Int> = mutableListOf()
 }
